extended the New... action
it starts with a wizard now which will get the Institute name, comment,
days, hours from the user.
After that, the old procedure comes: the selection of the extensions,
then opening the selected extension's wizard and giving them the
information gained from the new first wizard.
Not complete yet, it needs wizardpages.